Add files via upload
Laptop having issues screenshot from a video taken January 16th, 2025, and what caused me to upload all my personal and laptop-local class projects to GitHub in case it crashed. Luckily, it didn't, but until May, I hadn't had the chance to let campus IT backup files and fix this issue that happened for a few Zooms, as well as a touch pad bug.